pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etf-analyzer"
version = "0.1.0"
description = "ETF Fund Portfolio Investment Analysis System"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "akshare>=1.12",
    "matplotlib>=3.7",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools.packages.find]
where = ["src"]
